refactor: useTabsScroll 훅 분리
